requirements-deploy.txt
# Core ML Runtime
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
mlflow[databricks]==2.8.1

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1

# Testing (Optional but recommended)
pytest==7.4.4
httpx==0.26.0
